clean online resources

+++ scripts/clean-online.sh
@@ -10,49 +10,49 @@ echo "################ start clean EFS resource ################ "
 
 #delete EFS file system
 EFS_ID=$(aws efs describe-file-systems | jq '.[][] | select(.Tags[].Value=="GCR-RS-WORKSHOP-EFS-FileSystem")' | jq '.FileSystemId' -r)
-
-MOUNT_TARGET_IDS=$(aws efs describe-mount-targets --file-system-id $EFS_ID | jq '.[][].MountTargetId' -r)
-for MOUNT_TARGET_ID in `echo $MOUNT_TARGET_IDS`
-do
-  echo remove $MOUNT_TARGET_ID
-  aws efs delete-mount-target --mount-target-id $MOUNT_TARGET_ID
-done
-
-MOUNT_TARGET_IDS=""
-while true
-do
+if [ "$EFS_ID" != "" ]; then
   MOUNT_TARGET_IDS=$(aws efs describe-mount-targets --file-system-id $EFS_ID | jq '.[][].MountTargetId' -r)
-  if [ "$MOUNT_TARGET_IDS" == "" ];then
-    echo "delete GCR-RS-WORKSHOP-EFS-FileSystem EFS mount target successfully!"
-    break
-  else
-    echo "wait for GCR-RS-WORKSHOP-EFS-FileSystem EFS mount target deleted!"    
-  fi
-  sleep 10
-done
+  for MOUNT_TARGET_ID in $(echo $MOUNT_TARGET_IDS); do
+    echo remove $MOUNT_TARGET_ID
+    aws efs delete-mount-target --mount-target-id $MOUNT_TARGET_ID
+  done
 
-echo remove EFS File System: $EFS_ID
+  MOUNT_TARGET_IDS=""
+  while true; do
+    MOUNT_TARGET_IDS=$(aws efs describe-mount-targets --file-system-id $EFS_ID | jq '.[][].MountTargetId' -r)
+    if [ "$MOUNT_TARGET_IDS" == "" ]; then
+      echo "delete GCR-RS-WORKSHOP-EFS-FileSystem EFS mount target successfully!"
+      break
+    else
+      echo "wait for GCR-RS-WORKSHOP-EFS-FileSystem EFS mount target deleted!"
+    fi
+    sleep 10
+  done
 
-aws efs delete-file-system --file-system-id $EFS_ID
+  echo remove EFS File System: $EFS_ID
 
-EFS_ID=""
-while true
-do
-  EFS_ID=$(aws efs describe-file-systems | jq '.[][] | select(.Tags[].Value=="GCR-RS-WORKSHOP-EFS-FileSystem")' | jq '.FileSystemId' -r)
-  if [ "$EFS_ID" == "" ];then
-    echo "delete GCR-RS-WORKSHOP-EFS-FileSystem EFS successfully!"
-    break
-  else
-    echo "wait for GCR-RS-WORKSHOP-EFS-FileSystem EFS deleted!"    
-  fi
-  sleep 10
-done
+  aws efs delete-file-system --file-system-id $EFS_ID
+
+  EFS_ID=""
+  while true; do
+    EFS_ID=$(aws efs describe-file-systems | jq '.[][] | select(.Tags[].Value=="GCR-RS-WORKSHOP-EFS-FileSystem")' | jq '.FileSystemId' -r)
+    if [ "$EFS_ID" == "" ]; then
+      echo "delete GCR-RS-WORKSHOP-EFS-FileSystem EFS successfully!"
+      break
+    else
+      echo "wait for GCR-RS-WORKSHOP-EFS-FileSystem EFS deleted!"
+    fi
+    sleep 10
+  done
+fi
 
 #delete EFS file system security group
 NFS_SECURITY_GROUP_ID=$(aws ec2 describe-security-groups --filters Name=vpc-id,Values=$EKS_VPC_ID Name=group-name,Values=gcr-rs-workshop-efs-nfs-sg | jq '.SecurityGroups[].GroupId' -r)
-echo remove security group $NFS_SECURITY_GROUP_ID
 
-aws ec2 delete-security-group --group-id $NFS_SECURITY_GROUP_ID
+if [ "$NFS_SECURITY_GROUP_ID" != "" ]; then
+  echo remove security group $NFS_SECURITY_GROUP_ID
+  aws ec2 delete-security-group --group-id $NFS_SECURITY_GROUP_ID
+fi
 
 # i=1
 # NFS_SECURITY_GROUP_ID=""
@@ -63,7 +63,7 @@ aws ec2 delete-security-group --group-id $NFS_SECURITY_GROUP_ID
 #     echo "delete gcr-rs-workshop-efs-nfs-sg security group successfully!"
 #     break
 #   else
-#     echo "wait for gcr-rs-workshop-efs-nfs-sg security group deleted!"    
+#     echo "wait for gcr-rs-workshop-efs-nfs-sg security group deleted!"
 #   fi
 #   sleep 10
 #   let i++
@@ -75,14 +75,13 @@ echo "################ start Elasticache Redis resources ################ "
 aws elasticache delete-cache-cluster --cache-cluster-id gcr-rs-workshop-redis-cluster
 i=1
 ELASTIC_CACHE_CLUSTER=""
-while [ $i -le 20 ]
-do
+while [ $i -le 20 ]; do
   ELASTIC_CACHE_CLUSTER=$(aws elasticache describe-cache-clusters | jq '.CacheClusters[] | select(.CacheClusterId=="gcr-rs-workshop-redis-cluster")')
-  if [ "$ELASTIC_CACHE_CLUSTER" == "" ];then
+  if [ "$ELASTIC_CACHE_CLUSTER" == "" ]; then
     echo "delete gcr-rs-workshop-redis-cluster successfully!"
     break
   else
-    echo "wait for gcr-rs-workshop-redis-cluster deleted!"    
+    echo "wait for gcr-rs-workshop-redis-cluster deleted!"
   fi
   sleep 10
   let i++
@@ -105,7 +104,7 @@ aws ec2 delete-security-group --group-id $REDIS_SECURITY_GROUP_ID
 #     echo "delete gcr-rs-workshop-efs-nfs-sg security group successfully!"
 #     break
 #   else
-#     echo "wait for gcr-rs-workshop-efs-nfs-sg security group deleted!"    
+#     echo "wait for gcr-rs-workshop-efs-nfs-sg security group deleted!"
 #   fi
 #   sleep 10
 #   let i++
@@ -118,19 +117,10 @@ aws ec2 delete-security-group --group-id $REDIS_SECURITY_GROUP_ID
 echo "################ start Elasticache Redis resources ################ "
 
 ROLE_NAMES=$(aws iam list-roles | jq '.[][].RoleName' -r | grep eksctl-gcr-rs-workshop-cluster*)
-for ROLE_NAME in `echo $ROLE_NAMES`
-do
+for ROLE_NAME in $(echo $ROLE_NAMES); do
   echo remove $ROLE_NAME
   aws iam delete-role --role-name $ROLE_NAME
 done
 
 #remove eks cluster
 eksctl delete cluster --name=$EKS_CLUSTER
-
-
-
-
-
-
-
-
